Adjust the textbox poe count to be handled as a uint32_t

+++ GameCube/source/game_patch/05_itemMsgFunctions.cpp
@@ -167,7 +167,7 @@ namespace mod::game_patch
 
         // The poe count doesn't update until after the texbox has closed, so add one to get the new count
         uint32_t poeCount = libtp::tp::d_com_inf_game::dComIfG_gameInfo.save.save_file.player.player_collect.poe_count + 1;
-        return createString( format, msgSize, static_cast<uint8_t>( poeCount ) );
+        return createString( format, msgSize, poeCount );
     };
 
 #ifdef TP_EU
